+++ cmd/install.go
@@ -1,0 +1,23 @@
+package cmd
+
+import (
+	"github.com/spf13/cobra"
+	"streamres/installer"
+)
+
+// installCmd represents the install command
+var installCmd = &cobra.Command{
+	Use:   "install",
+	Short: "Installs and configures sunshine to use streamres",
+	Long: `Installs streamres by creating configuration directories and moving some files. 
+This includes installing dependencies and setting up a cache folder, 
+copying streamres to the sunshine tools folder, 
+and setting up system startup and shutdown scripts to allow clean recovery of monitor configuration`,
+	RunE: func(cmd *cobra.Command, args []string) error {
+		return installer.Install()
+	},
+}
+
+func init() {
+	rootCmd.AddCommand(installCmd)
+}