+++ cmd/start.go
@@ -4,7 +4,9 @@ import (
 	"fmt"
 	"github.com/spf13/cobra"
 	"streamres/displays"
+	"streamres/validate"
 	"strings"
+	"time"
 )
 
 var (
@@ -19,6 +21,9 @@ var startCmd = &cobra.Command{
 	Short: "Enables the virtual monitor and sets configuration",
 	Long: `start enables the virtual monitor and sets the resolution and refresh rate requested by the client.
 Disables any attached physical monitors while in use and stores the state of the system before starting to allow a revert`,
+	PreRunE: func(cmd *cobra.Command, args []string) error {
+		return validate.Application()
+	},
 	RunE: func(cmd *cobra.Command, args []string) error {
 		// Get monitor state first
 		state := displays.GetCurrentState()
@@ -43,7 +48,10 @@ Disables any attached physical monitors while in use and stores the state of the
 		// Found virtual monitor. Enable, set the resolution and mark it as the primary monitor.
 		displays.Enable(*virtualDisplay)
 		displays.SetConfig(displays.DisplayConfig{Width: width, Height: height, RefreshRate: refreshRate}, *virtualDisplay)
+		// Wait for changes to settle otherwise monitors can end up flickering
+		time.Sleep(1 * time.Second)
 		displays.SetAsPrimary(*virtualDisplay)
+		time.Sleep(1 * time.Second)
 
 		// Finally turn off all other monitors
 		for _, display := range otherDisplays {